largest element in an array
In the above program, we store the first element of the array in the variable largest.
Then, largest is used to compare other elements in the array. If any number is greater than largest, largest is assigned the number.
In this way, the largest number is stored in largest when it is printed.